chore: run build on variety of operating systems
