+++ src/it/test-with-args/pom.xml
@@ -4,7 +4,7 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>io.snyk.it</groupId>
-  <artifactId>test-without-dependency</artifactId>
+  <artifactId>test-with-args</artifactId>
   <version>1.0-SNAPSHOT</version>
 
   <build>
@@ -23,7 +23,7 @@
         </executions>
         <configuration>
           <cli>
-            <executable>/usr/local/bin/snyk</executable>
+            <executable>${env.SNYK_CLI_EXECUTABLE}</executable>
           </cli>
           <args>
             <arg>--print-deps</arg>
