+++ src/main/java/io/snyk/snyk_maven_plugin/AbstractSnykMojo.java
@@ -1,25 +1,22 @@
 package io.snyk.snyk_maven_plugin;
 
-import com.google.common.annotations.VisibleForTesting;
 import io.snyk.snyk_maven_plugin.command.Command;
 import io.snyk.snyk_maven_plugin.command.CommandLine;
 import io.snyk.snyk_maven_plugin.command.CommandRunner;
-import io.snyk.snyk_maven_plugin.download.GitHubDownloader;
+import io.snyk.snyk_maven_plugin.download.ExecutableDownloader;
 import io.snyk.snyk_maven_plugin.download.Installer;
 import io.snyk.snyk_maven_plugin.download.Platform;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
+import org.apache.maven.plugin.logging.Log;
 import org.apache.maven.plugins.annotations.Parameter;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.InvalidPathException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
-import java.util.Map;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
 public abstract class AbstractSnykMojo extends AbstractMojo {
@@ -52,7 +49,7 @@ public abstract class AbstractSnykMojo extends AbstractMojo {
                     try {
                         Platform platform = Platform.current();
                         Path targetFolder = Installer.getInstallLocation(platform, Optional.ofNullable(System.getProperty("user.home")).map(Paths::get), System.getenv());
-                        return GitHubDownloader.download(targetFolder, Platform.current());
+                        return ExecutableDownloader.download(targetFolder, Platform.current());
                     } catch (IOException e) {
                         throw new RuntimeException("failed to download executable", e);
                     }
